(refactor): Correct DI
